+++ DungeonsAndDatabases.WebAPI/Controllers/CharacterController.cs
@@ -22,9 +22,14 @@ namespace DungeonsAndDatabases.WebAPI.Controllers
         }
 
         //CREATE
+        /// <summary>
+        /// Create a Character
+        /// </summary>
+        /// <param name="cha"></param>
+        /// <returns>Creates a character and assigns it to the current player</returns>
         //Create New Character
         [HttpPost]
-        public async Task<IHttpActionResult> Character(CharacterCreate cha)
+        public async Task<IHttpActionResult> CreateCharacter(CharacterCreate cha)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
@@ -36,8 +41,11 @@ namespace DungeonsAndDatabases.WebAPI.Controllers
                 return InternalServerError();
             return Ok();
         }
-
         //READ
+        /// <summary>
+        /// Get all Characters
+        /// </summary>
+        /// <returns>Returns a list of Characters belonging to player</returns>
         //Get All Characters
         [HttpGet]
         public async Task<IHttpActionResult> GetCharacter()
@@ -48,6 +56,11 @@ namespace DungeonsAndDatabases.WebAPI.Controllers
         }
 
         //Get Character By ID
+        /// <summary>
+        /// Get Character By ID
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns>Gets a specific character by it's ID</returns>
         [HttpGet]
         public async Task<IHttpActionResult> GetCharacter(int id)
         {
@@ -57,29 +70,43 @@ namespace DungeonsAndDatabases.WebAPI.Controllers
         }
 
         //Update a Character
+        /// <summary>
+        /// Update a Character
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="character"></param>
+        /// <returns>Updates a character that the player owns</returns>
         [HttpPut]
         public async Task<IHttpActionResult> UpdateCharacter([FromUri] int id, [FromBody] CharacterEdit character)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
             var service = CreateCharacterService();
-
+            var credentials = await service.CheckCharCredentials(id);
+            if (credentials == false)
+                return Unauthorized();
             var result = await service.UpdateCharacter(id, character);
-
-            if (!service.UpdateCharacter(id, character).Result)
+            if (result == false)
                 return InternalServerError();
             return Ok();
         }
 
         //Delete a Character
+        /// <summary>
+        /// Delete a Character
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns>Delete a character that a player owns</returns>
         [HttpDelete]
         public async Task<IHttpActionResult> DeleteCharacter(int id)
         {
             var service = CreateCharacterService();
-
+            var credentials = await service.CheckCharCredentials(id);
+            if (credentials == false)
+                return Unauthorized();
             var result = await service.DeleteCharacter(id);
 
-            if (!service.DeleteCharacter(id).Result)
+            if (result == false)
                 return InternalServerError();
             return Ok();
         }
